AgenticRAG/pyproject.toml
[project]
name = "agenticrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.4.22",
    "docx2txt>=0.8",
    "google-adk>=1.22.1",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-text-splitters>=1.1.0",
    "litellm>=1.20.0",
    "opik>=1.9.87",
    "pypdf>=3.17.4",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.3.1",
    "streamlit>=1.30.0",
]
